+++ README.md
@@ -1,2 +1,4 @@
-# northstar-node-client
-Node.js client for Northstar, the DoSomething.org user API.
+# northstar-js
+Javascript client for Northstar, the DoSomething.org user API.
+
+Currently this is Node.js implementation only.
