+++ package.json
@@ -1,25 +1,25 @@
 {
-  "name": "northstar-node-client",
+  "name": "northstar-js",
   "version": "0.0.0",
-  "description": "Node.js client for Northstar, the DoSomething.org user API.",
+  "description": "Javascript client for Northstar, the DoSomething.org user API.",
   "engines": {
     "node": ">=4.4.x"
   },
-  "main": "./lib/northstar-node-client",
+  "main": "./lib/northstar-client",
   "scripts": {
     "test": "npm run lint && mocha",
     "lint": "eslint test/*.js lib/*.js"
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/DoSomething/northstar-node-client.git"
+    "url": "git+https://github.com/DoSomething/northstar-js.git"
   },
   "author": "Sergii Tkachenko <hi@sergii.org>",
   "license": "MIT",
   "bugs": {
-    "url": "https://github.com/DoSomething/northstar-node-client/issues"
+    "url": "https://github.com/DoSomething/northstar-js/issues"
   },
-  "homepage": "https://github.com/DoSomething/northstar-node-client#readme",
+  "homepage": "https://github.com/DoSomething/northstar-js#readme",
   "devDependencies": {
     "@dosomething/eslint-config": "^1.1.1",
     "eslint": "^2.10.2",
